editor/libeditor/tests/test_bug1153237.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1153237
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 1153237</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript">

  SimpleTest.waitForExplicitFinish();

  // Avoid platform selection differences
  SimpleTest.waitForFocus(function() {
    SpecialPowers.pushPrefEnv({
      "set": [["layout.word_select.eat_space_to_next_word", true]]
    }, runTests);
  });

  function runTests()
  {
    var element = document.getElementById("editor");
    var sel = window.getSelection();

    element.focus();
    is(sel.getRangeAt(0).startOffset, 0, "offset is zero");

    SpecialPowers.doCommand(window, "cmd_selectRight2");
    is(sel.toString(), "Some ",
       "first word + space is selected: got '" + sel.toString() + "'");

    SpecialPowers.doCommand(window, "cmd_selectRight2");
    is(sel.toString(), "Some text",
       "both words are selected: got '" + sel.toString() + "'");

    SimpleTest.finish();
  }
  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1153237">Mozilla Bug 1153237</a>
<div id="editor" contenteditable>Some text</div><span></span>

<pre id="test">
</pre>
</body>
</html>
